requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
sqlalchemy>=2.0.29,<3.0.0
pydantic>=2.4.0,<3.0.0
pydantic-settings>=2.2.1,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
python-multipart>=0.0.9,<1.0.0
pytest>=8.2.0,<9.0.0
httpx>=0.27.0,<1.0.0
requests>=2.31.0,<3.0.0
PyQt5>=5.15.10,<6.0.0
